Create `winnet` module
Based on the previous `route` module.
