+++ talpid-core/src/security/windows/dns.rs
@@ -1,16 +1,14 @@
-extern crate widestring;
-
-use super::ffi;
-use super::system_state::SystemStateWriter;
-
-use self::widestring::WideCString;
-use libc;
 use std::net::IpAddr;
 use std::path::Path;
 use std::ptr;
 use std::slice;
 
 use error_chain::ChainedError;
+use libc;
+use widestring::WideCString;
+
+use super::system_state::SystemStateWriter;
+use winnet;
 
 const DNS_STATE_FILENAME: &'static str = "dns-state-backup";
 
@@ -49,7 +47,7 @@ pub struct WinDns {
 
 impl WinDns {
     pub fn new<P: AsRef<Path>>(cache_dir: P) -> Result<Self> {
-        unsafe { WinDns_Initialize(Some(ffi::error_sink), ptr::null_mut()).into_result()? };
+        unsafe { WinDns_Initialize(Some(winnet::error_sink), ptr::null_mut()).into_result()? };
 
         let backup_writer = SystemStateWriter::new(
             cache_dir
@@ -191,7 +189,7 @@ extern "system" {
 
     #[link_name(WinDns_Initialize)]
     pub fn WinDns_Initialize(
-        sink: Option<ffi::ErrorSink>,
+        sink: Option<winnet::ErrorSink>,
         sink_context: *mut libc::c_void,
     ) -> InitializationResult;
