+++ talpid-core/src/winnet.rs
@@ -1,7 +1,7 @@
-use super::ffi;
-use super::widestring::WideCString;
-use libc;
 use std::ptr;
+
+use libc::{c_char, c_void, wchar_t};
+use widestring::WideCString;
 
 error_chain!{
     errors{
@@ -15,6 +15,17 @@ error_chain!{
     }
 }
 
+pub type ErrorSink = extern "system" fn(msg: *const c_char, ctx: *mut c_void);
+
+pub extern "system" fn error_sink(msg: *const c_char, _ctx: *mut c_void) {
+    use std::ffi::CStr;
+    if msg.is_null() {
+        error!("Log message from FFI boundary is NULL");
+    } else {
+        error!("{}", unsafe { CStr::from_ptr(msg).to_string_lossy() });
+    }
+}
+
 /// Returns true if metrics were changed, false otherwise
 pub fn ensure_top_metric_for_interface(interface_alias: &str) -> Result<bool> {
     let interface_alias_ws =
@@ -22,7 +33,7 @@ pub fn ensure_top_metric_for_interface(interface_alias: &str) -> Result<bool> {
     unsafe {
         WinRoute_EnsureTopMetric(
             interface_alias_ws.as_wide_c_str().as_ptr(),
-            Some(ffi::error_sink),
+            Some(error_sink),
             ptr::null_mut(),
         ).into()
     }
@@ -56,8 +67,8 @@ impl Into<Result<bool>> for MetricResult {
 extern "system" {
     #[link_name(WinRoute_EnsureTopMetric)]
     fn WinRoute_EnsureTopMetric(
-        tunnel_interface_alias: *const libc::wchar_t,
-        sink: Option<ffi::ErrorSink>,
-        sink_context: *mut libc::c_void,
+        tunnel_interface_alias: *const wchar_t,
+        sink: Option<ErrorSink>,
+        sink_context: *mut c_void,
     ) -> MetricResult;
 }
